koniec spotkania

+++ P/Zadania/R1/R1.2/zad11.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    double x;
+    long double x;
     cout<<"Podaj liczbe wymierna: "; cin>>x;
     cout<<x<<endl;
 }
